requirements.txt
streamlit
PyPDF2
pdf2docx
PyMuPDF
pdf2image
Pillow
python-pptx
openpyxl
ebooklib
pytesseract
transformers
torch
sentencepiece
asyncio
whoosh
python-magic
Babel
gunicorn
uvicorn
numpy
scipy
